とりあえず型をtypes以下に移動
